requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pydantic
python-multipart
google-cloud-discoveryengine